refactor cooperative signaling folder structure
