+++ abm/projects/cooperative_signaling/app.py
@@ -5,8 +5,7 @@ from pathlib import Path
 from dotenv import dotenv_values
 
 from abm.app import save_isims_env
-from simulation import CSPlaygroundSimulation, \
-    setup_coop_sign_playground
+from cs_simulation.cs_isims import CSPlaygroundSimulation
 
 
 def setup_environment():
@@ -27,6 +26,8 @@ def start_playground():
     env_file_dir, EXP_NAME_COPY, envconf = setup_environment()
     # changing env file according to playground default parameters before
     # running any component of the SW
+    from abm.projects.cooperative_signaling.contrib.cs_playgroundtool import \
+        setup_coop_sign_playground
     pgt = setup_coop_sign_playground()
     save_isims_env(env_file_dir, EXP_NAME_COPY, pgt, envconf)
     # Start interactive simulation
